Update AdditivePersistence.java
Using longs just extends the range of values that can be input into the program.

+++ AdditivePersistence.java
@@ -8,10 +8,10 @@ import java.util.Scanner;
 
 public class AdditivePersistence {
     /** This is the number the user will enter when prompted. */
-    private static int userNum;
+    private static long userNum;
 
     /** This is the number that will be used to sum the individual terms of the user's number. */
-    private static int sums;
+    private static long sums;
 
     /** This counter will keep track of how many times we added terms to get to a single term number. */
     private static int counter = 1;
@@ -22,7 +22,7 @@ public class AdditivePersistence {
         Scanner scnr = new Scanner(System.in);
 
         System.out.print("Please enter your number: ");
-        userNum = scnr.nextInt();
+        userNum = scnr.nextLong();
 
         AddPersistence(userNum);
 
@@ -33,9 +33,9 @@ public class AdditivePersistence {
     /** Does the heavy lifting by adding individual terms of a given number
     and doing so until it is a single term number (0-9). 
     * @return none
-    * @param int num - the number whose additive persistence we want to compute.
+    * @param long num - the number whose additive persistence we want to compute.
     */
-    public static void AddPersistence(int num) {
+    public static void AddPersistence(long num) {
         while(num > 0) {
             sums += num % 10;
             num = num / 10;
